+++ crates/event-bus-dto/schemas/events.json
@@ -183,7 +183,7 @@
         "type": "object"
       },
       "QuoteRequestedEvent": {
-        "description": "Emitted when the orderbook starts computing a quote (\"calculating quote\"),\nbefore any price estimation runs. Carries the request context that partner\nand quote analysis care about — notably the `appCode` and the token symbols\n— which are not present on the per-estimator [`crate::PriceEstimateEvent`].",
+        "description": "Emitted for a validated quote request, just before price estimation runs.\nCarries the request context that partner and quote analysis care about —\nnotably the `appCode` and the token symbols — which are not present on the\nper-estimator [`crate::PriceEstimateEvent`].",
         "properties": {
           "appCode": {
             "description": "`appCode` from the order's app-data document, if present.",
