+++ crates/event-bus-dto/src/quote_requested.rs
@@ -1,9 +1,13 @@
-use {crate::Event, schemars::JsonSchema, serde::Serialize};
+use {
+    crate::{Event, query::QueryFields},
+    schemars::JsonSchema,
+    serde::Serialize,
+};
 
-/// Emitted when the orderbook starts computing a quote ("calculating quote"),
-/// before any price estimation runs. Carries the request context that partner
-/// and quote analysis care about — notably the `appCode` and the token symbols
-/// — which are not present on the per-estimator [`crate::PriceEstimateEvent`].
+/// Emitted for a validated quote request, just before price estimation runs.
+/// Carries the request context that partner and quote analysis care about —
+/// notably the `appCode` and the token symbols — which are not present on the
+/// per-estimator [`crate::PriceEstimateEvent`].
 #[derive(Serialize, JsonSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct QuoteRequestedEvent {
@@ -27,25 +31,6 @@ impl Event for QuoteRequestedEvent {
 }
 
 #[derive(Serialize, JsonSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryFields {
-    /// Hex-encoded sell token address.
-    pub sell_token: String,
-    /// Hex-encoded buy token address.
-    pub buy_token: String,
-    /// Decimal-encoded input amount (interpretation depends on `kind`).
-    pub in_amount: String,
-    pub kind: OrderKind,
-}
-
-#[derive(Serialize, JsonSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum OrderKind {
-    Sell,
-    Buy,
-}
-
-#[derive(Serialize, JsonSchema)]
 #[serde(rename_all = "lowercase")]
 pub enum PriceQuality {
     Fast,
@@ -55,7 +40,7 @@ pub enum PriceQuality {
 
 #[cfg(test)]
 mod tests {
-    use {super::*, serde_json::json};
+    use {super::*, crate::query::OrderKind, serde_json::json};
 
     #[test]
     fn matches_wire_format() {
